feat(feature/#20): 이미지 업로드 부분 컴포넌트 분리 및 image 업데이트
